omapdss: displays: Add tfcs900 panel support
Add a panel driver for tfcs9000. This isn't a correct implementation of the
driver, as it also acts as an i2c driver for the tlc59108 pwm/gpio expander
driver.

The tlc chip is configured to make sure only the pwm and gpios needed by the
panel are configured. The rest are untouched.

Signed-off-by: Archit Taneja <archit@ti.com>