omapdss: DPI: Use PLL library
Make the DPI driver use the pll helper library to use DSI PLLs to compute the
PLL parameters for the desired pixel clock, and to configure and lock the PLL.

DPI in itself doesn't create PLL instances, it retrieves pll_data from the DSI
platform device. The driver still relies on the dsi functions to enable/disable
the DSI PLLs. This would later be removed when we add PLL ops to hide this from
the DPI driver.

Signed-off-by: Archit Taneja <archit@ti.com>