OMAPDSS: PLL: Add pll ops
PLL parameter computation and configuration is generally independent of the IP
in which the PLL belongs to. But power and clock control of the PLL is managed
by the DSI protocol engine, and HDMI wrapper in DSI and HDMI respectively.

In order to make the DPI driver completely independent of DSI, we need to add
ops in the PLL library which the DSI(or HDMI) driver populates. These ops will
performs the power, clock and regulator management related operations managed
by the parent IP.

Signed-off-by: Archit Taneja <archit@ti.com>