drm/msm/mdp5: Add START signal to kick off certain pipelines
Some interfaces (WB, DSI Command Mode) need to be kicked off
through a START Signal. This signal needs to be sent at the right
time and requests in some cases to keep track of the pipeline
status (eg: whether pipeline registers are flushed AND output WB
buffers are ready, in case of WB interface).

Change-Id: Ic2ccc6470cc6861954df775b68caa33e4bcb4af2
Signed-off-by: Stephane Viau <sviau@codeaurora.org>