drm: adv7511: Configure as a bridge driver
ADV75xx(especially ADV7533) can be represented as a bridge driver rather than
an i2c slave. This is because it expects DSI data as input, which is generally
the job of an encoder.

Add a config option which lets us chose between normal i2c slave encoder and
bridge usage. DRM_I2C_ADV7511_SLAVE_ENCODER is set by default, making adv75xx
behave as a slave encoder. Setting it to 'n' makes the driver create a bridge
and hdmi connector entities.

Signed-off-by: Archit Taneja <architt@codeaurora.org>