drm/i2c: adv7511: Create mipi_dsi_device for ADV7533
In order to pass DSI specific parameters to the DSI host, we need the
driver to create a mipi_dsi_device that attaches to the host.

Parse the DT phandle to get the DSI host. Create a dummy dsi device using
this host. Finally, attach this device to the host.

Populate few other DT parameters (number of data lanes etc) that are
required for DSI RX to work correctly. Hardcode few other parameters
(rgb, embedded_sync) for now.

Signed-off-by: Archit Taneja <architt@codeaurora.org>