drm/i2c: adv7511: Add dsi driver for adv7533
Add a separate mipi_dsi_driver for adv7533. In the case of this chip,
both the i2c and dsi drivers will operate together.

Both the drivers are expected to use the same per-device driver
data struct. The i2c driver takes the responsibility of allocating
the struct, and the dsi device gets a pointer to it by getting the
corresponding i2c client device's data.

Signed-off-by: Archit Taneja <architt@codeaurora.org>