drm/msm/mdp5: get additional mmagic ahb/iommu clocks for MSM8996
In order to access MDP registers on MSM8996, we need to enable an
additional AHB clock called MMAGIC AHB clock. We also need to enable a
MDP SMMU AXI clock to fetch frame buffers, even when we are are bypassing
the IOMMU. Retrieve these clocks as optional clocks for now.

Signed-off-by: Archit Taneja <architt@codeaurora.org>