drm/msm/mdp5: add a separate MDP5 device
In order to have a tree-like device hierarchy between MDSS and its
sub-blocks (MDP5, DSI, HDMI, eDP etc), we need to create a separate
device/driver for MDP5. Currently, it's tied to drm_device struct's
platform device.

The mdp5_kms_init code is split into two parts. The part where device
resources are allocated are associated with the MDP5 driver's probe, and
the rest is executed later when we initialize modeset.

With this change, unlike MDP4, the MDP5 platform_device isn't tied to
the top level drm_device anymore. The top level drm_device is now
associated with a platform device that corresponds to MDSS wrapper
hardware.

Create mdp5_init func that will be used when the MDP5 driver probes.
Use the HW_VERSION register in the MDP5 register address space.

The older mdp5_kms_init code is left as is for now, this would be
removed later when we have all the pieces to support the new
hierarchy.

Signed-off-by: Archit Taneja <architt@codeaurora.org>