drm/msm/mdp5: Prepare new kms_init funcs
With MDP5 as a new device, we need to do less for MDP when initializing
modeset after all the components are bound.

Create mdp5_kms_init2/destroy2 funcs that inits modeset. These will
eventually replace the older kms_init/destroy funcs.

In the new kms_init2, the platform_device used is the one corresponding
to the new MDP5 platform_device. The new changes in this func are:

 1) The MDSS helper library is also initialized (which results in setting
    up the top level MDSS device resources).
 2) The irq is now retrieved using irq_of_parse_and_map(), since MDP5 is
    a child interrupt of the MDSS interrupt controller.

Signed-off-by: Archit Taneja <architt@codeaurora.org>