drm/msm/mdp5: introduce new bindings for MDSS/MDP5
Introduce the new bindings for the top level MDSS wrapper device,
and MDP5.

Previously, the "qcom,mdp5" and "qcom,mdss_mdp" compatible strings
were used to match the top level platform_device (that is also tied
to the top level drm_device struct). Now, these strings are used
to match the MDP5 platform device, which is a child of the MDSS
device.

Use "qcom,mdss" as the compatible string for top level MDSS device.

Signed-off-by: Archit Taneja <architt@codeaurora.org>