dt-bindings: drm/bridge: adv7511: Add regulator bindings
Add the regulator supply properties needed for ADV7533. The regulator
names are named after the pin names. The exact use of each supply isn't
clear from the data sheets.

The regulators are specified as optional properties since there can
be boards which have a fixed supply directly routed to the pins, and
these may not be modelled as regulator supplies.

ADV7511 has a BGVDD pin for which the voltage to be configured isn't
clear. This needs to be updated in the document later.

Cc: Rob Herring <robh@kernel.org>
Cc: devicetree@vger.kernel.org
Signed-off-by: Archit Taneja <architt@codeaurora.org>