drm/msm: Set encoder's mode of operation using a kms func
The mdp5 kms driver currently sets up multiple multiple encoders per
interface (INTF), one for each kind of mode of operation it supports.
In particular, we create 2 drm_encoders for DSI, one for video mode
and the other for command mode operation. The reason behind this
approach could be that we aren't aware of the DSI device's mode
of operation when we create the encoders.

This makes things a bit complicated, since these encoders have to
be further attached to the same DSI bridge. The easier way out is
to create a single encoder, and make the DSI driver set its mode
of operation when we know what the DSI device's mode flags are.

Start with providing a way to set the mdp5_intf_mode using a kms
func. Later, we remove the multiple encoders created for DSI.

Signed-off-by: Archit Taneja <architt@codeaurora.org>