drm/msm/mdp5: Prepare CRTC/LM for empty stages
Use SSPP_NONE in mdp5_plane_pipe() if there is now hwpipe allocated for
the drm_plane. Returning '0' means we are returning VIG0 pipe.

Also, use the mdp5_pipe enum to pass around the stage array. Initialize
the stage to SSPP_NONE by default.

We do the above because 1) Cursor plane is staged at the topmost blender
of the LM 2) In the future, when we support multiple LMs per CRTC. We
could have stages which don't have any pipe assigned to them.

Signed-off-by: Archit Taneja <architt@codeaurora.org>