drm/msm: Use dma_sync_sg_for_dev to clean caches
The msm gem driver, like many other drm drivers, doesn't rely on the dma
mapping API to manage the IOMMU domains for the display and gpu
hardware. We do our IOMMU domain/IOVA management ourselves.

We currently use dma_map_sg() on our gem buffer to flush he caches for the
display/gpu device to have a consistent view of the buffer. This works
fine since our device's dma_map_ops are currently set to the swiotlb ops.
This could become a problem if the dma_map_ops are set to iommu ops in the
future, which will result in both the dma mapping API and the gem driver
try to map the buffer to the IOMMU domain.

Use the dma_sync_sg_for_device() dma mapping API, whose job is to only
synchronize the buffer for the device.

When free-ing the page, the driver currently calls dma_unmap_sg() on
the buffer sg list. We don't need to do this since we're free-ing the
page anyway.

The code is borrowed from the tegra/rockchip gem drivers.

Signed-off-by: Archit Taneja <architt@codeaurora.org>

squash with cache